migrated search out of app component to its own page and fixed routing
