+++ demo/smallbusiness-ui/src/app/search/search.component.html
@@ -1,1 +1,52 @@
-<button (click)="getLocalBusinesses(43.157109, -77.615179, 'coffee')">Test API Request</button>
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <meta http-equiv="X-UA-Compatible" content="ie=edge">
+  <title>Home</title>
+  <link href="search.component.css" rel="stylesheet">
+  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-Zenh87qX5JnK2Jl0vWa8Ck2rdkQ2Bzep5IDxbcnCeuOxjzrPF/et3URy9Bv1WTRi" crossorigin="anonymous">
+
+  <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-OERcA2EqjJCMA+/3y+gxIOqMEjwtxJY7qPCqsdltbNJuaOe923+mo//f6V8Qbsw3" crossorigin="anonymous"></script>
+
+
+</head>
+
+<body>
+
+  <!-- NAVBAR -->
+  <nav class="navbar navbar-expand-sm bg-dark">
+    <div class="container-fluid">
+      <ul class="navbar-nav">
+        <li class="nav-item">
+          <a class="nav-link text-light" routerLink="">Home</a>
+        </li>
+        <li class="nav-item">
+          <a class="nav-link text-light" routerLink="/about-page">About</a>
+        </li>
+        <li class="nav-item">
+          <a class="nav-link text-light" routerLink="/contact-us">Contact Us</a>
+        </li>
+      </ul>
+    </div>
+  </nav>
+  <router-outlet></router-outlet>
+
+  <h1>Hello World</h1>
+
+  <button (click)="getLocalBusinesses(40.333708, -75.637408, 'coffee')">Search</button>
+
+  <div *ngFor="let business of localCompanies">
+    <div class="businessView">
+      {{business.name}}
+    </div>
+  </div>
+
+
+
+</body>
+
+
+</html>
